+++ src/OFX.SDK/Primitives/MessageSets/IOFXCancelMessage.cs
