+++ src/OFX.SDK/Primitives/MessageSets/IOFXDeleteMessage.cs
