+++ src/OFX.SDK/Primitives/MessageSets/IOFXInquiryMessage.cs
